Adding an integer generated feature type to FieldProcessor.

+++ Data/src/main/java/org/tribuo/data/columnar/processors/field/DateFieldProcessor.java
@@ -208,7 +208,7 @@ public final class DateFieldProcessor implements FieldProcessor {
 
     @Override
     public GeneratedFeatureType getFeatureType() {
-        return GeneratedFeatureType.REAL;
+        return GeneratedFeatureType.INTEGER;
     }
 
     @Override
